requirements.txt
granian==1.5.1
cassandra-driver==3.29.1
fastapi-slim==0.111.0
python-json-logger==2.0.7
concurrent-log-handler==0.9.25
asgi-logger==0.1.0
pydantic==2.8.2
pydantic-settings==2.4.0
PyYAML==6.0.1
python-multipart==0.0.9

requests==2.32.3
httpx==0.27.0
urllib3==2.2.2
tenacity==9.0.0

setuptools==72.1.0